moved duplicate functions to utils file
